refactor(featureMatching): create subfolders
