+++ src/featureMatching/descriptors/__tests__/__image_snapshots__/get-brief-descriptors-test-ts-alphabet-image-max-nb-features-10-1-snap.png
