+++ src/featureMatching/descriptors/__tests__/__image_snapshots__/get-brief-descriptors-test-ts-should-work-with-small-patch-size-1-snap.png
