+++ src/featureMatching/draw/__tests__/__image_snapshots__/draw-matches-test-ts-alphabet-image-as-source-and-destination-nb-keypoint-10-1-snap.png
