+++ src/featureMatching/draw/__tests__/__image_snapshots__/draw-matches-test-ts-destination-rotated-2-1-snap.png
