+++ src/featureMatching/index.ts
@@ -1,7 +1,7 @@
-export * from './bruteForceMatch';
-export * from './drawKeypoints';
-export * from './drawMatches';
-export * from './getBriefDescriptors';
-export * from './getFastKeypoints';
-export * from './getIntensityCentroid';
-export * from './getOrientedFastKeypoints';
+export * from './matching/bruteForceMatch';
+export * from './draw/drawKeypoints';
+export * from './draw/drawMatches';
+export * from './descriptors/getBriefDescriptors';
+export * from './keypoints/getFastKeypoints';
+export * from './keypoints/getIntensityCentroid';
+export * from './keypoints/getOrientedFastKeypoints';
