+++ src/featureMatching/keypoints/__tests__/__image_snapshots__/get-fast-keypoints-test-ts-alphabet-fast-radius-5-1-snap.png
