+++ src/featureMatching/keypoints/__tests__/__image_snapshots__/get-fast-keypoints-test-ts-alphabet-image-default-options-1-snap.png
