+++ src/featureMatching/keypoints/__tests__/__image_snapshots__/get-fast-keypoints-test-ts-alphabet-image-score-algorithm-harris-max-nb-features-50-1-snap.png
