+++ src/featureMatching/keypoints/__tests__/__image_snapshots__/get-fast-keypoints-test-ts-alphabet-image-score-algorithm-harris-nms-false-normalize-scores-true-2-snap.png
