+++ src/featureMatching/keypoints/__tests__/__image_snapshots__/get-fast-keypoints-test-ts-grayscale-image-threshold-100-normalize-scores-true-2-snap.png
