+++ src/featureMatching/keypoints/__tests__/getHarrisScore.test.ts
@@ -1,4 +1,4 @@
-import { ImageColorModel, Image } from '../../Image';
+import { ImageColorModel, Image } from '../../../Image';
 import { getHarrisScore } from '../getHarrisScore';
 
 const fastRadius = 3;
