+++ src/featureMatching/keypoints/__tests__/getIntensityMoment.test.ts
@@ -1,4 +1,4 @@
-import { ImageColorModel, Image } from '../../Image';
+import { ImageColorModel, Image } from '../../../Image';
 import { getIntensityMoment } from '../getIntensityMoment';
 
 test('3x3 empty image, 00', () => {
