+++ src/featureMatching/keypoints/__tests__/getOrientedFastKeypoints.test.ts
