+++ src/featureMatching/matching/__tests__/getHammingDistance.test.ts
